+++ libs/@wingcloud/framework/test/fixtures/fail/unlifted_reference.ts
@@ -1,0 +1,9 @@
+import { main, inflight } from "../../../src";
+
+main((root) => {
+  let x = 2;
+
+  inflight(async () => {
+    console.log(x);
+  });
+});
