+++ libs/@wingcloud/framework/test/fixtures/pass/autolift_import.ts
@@ -1,0 +1,8 @@
+import { main, inflight } from "../../../src";
+import { equal } from "node:assert";
+
+main((root) => {
+  inflight(async () => {
+    equal(1, 1);
+  });
+});
