+++ libs/@wingcloud/framework/test/fixtures/pass/basic_inflight.ts
@@ -1,0 +1,7 @@
+import { main, inflight } from "../../../src";
+
+main((root) => {
+  inflight(async () => {
+    console.log("Hello, world!");
+  });
+});
